+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- newest-changes -->
+## [0.0.3] - 2015-12-22
+### Added
+    * --fork option to work with Github API
+    * option to prefer git@ over https://
+    * configurable clone opts
+    * CLI overrideable config
+### Fixed
+    * Simplified code to use $self instead of $location in general
+
 ## [0.0.2] - 2015-12-22
 ### Added
     * `show` command
@@ -19,5 +28,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Initial commit
 
 <!-- link-labels -->
+[0.0.3]: ../compare/v0.0.2...v0.0.3
 [0.0.2]: ../compare/v0.0.1...v0.0.2
 [0.0.1]: ../compare/v0.0.1...HEAD